+++ shardis-ui/config/webpack.commons.js
@@ -1,7 +1,8 @@
-var path = require('path');
-var autoprefixer = require('autoprefixer');
-var validateWebpackConfig = require('webpack-validator');
-var Joi = require('joi');
+const path = require('path');
+const autoprefixer = require('autoprefixer');
+const ForkCheckerPlugin = require('awesome-typescript-loader').ForkCheckerPlugin;
+const validateWebpackConfig = require('webpack-validator');
+const Joi = require('joi');
 
 // Environments
 const ENV = {
@@ -41,7 +42,7 @@ var baseWebpackConfig = {
 
   cache: true,
   debug: false,
-  
+
   tslint: {
     emitErrors: false,
     failOnHint: false,
@@ -71,7 +72,7 @@ var baseWebpackConfig = {
     loaders: [
       {
         test: /\.ts$/,
-        loader: 'ts'
+        loader: 'awesome-typescript-loader'
       },
 
       {
@@ -117,7 +118,9 @@ var baseWebpackConfig = {
   postcss: [
     autoprefixer({browsers: ['last 2 versions'], remove: false})
   ],
-  plugins: []
+  plugins: [
+    new ForkCheckerPlugin()
+  ]
 };
 
 module.exports = {